requirements.txt
Django~=3.2.13
pytz~=2022.1
djangorestframework~=3.13.1